add a bunch of function logging

+++ functions/EmuScripts/emuDeckCemu.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #variables
-Cemu_emuName="cemu"
+Cemu_emuName="Cemu"
 Cemu_emuType="windows"
 Cemu_emuPath="${romsPath}wiiu/cemu.exe"
 Cemu_releaseURL="https://cemu.info/releases/cemu_1.26.2.zip"
@@ -13,20 +13,16 @@ Cemu.cleanup(){
 
 #Install
 Cemu.install(){
-	setMSG "Installing Cemu"		
-	FILE="${romsPath}/wiiu/Cemu.exe"	
-	if [ -f "$FILE" ]; then
-		echo "Cemu.exe already exists"
-	else
-		curl $Cemu_releaseURL --output "$romsPath"wiiu/cemu.zip 
-        mkdir -p "$romsPath"wiiu/tmp
-        unzip -o "$romsPath"wiiu/cemu.zip -d "$romsPath"wiiu/tmp
-        mv "$romsPath"wiiu/tmp/cemu_*/ "$romsPath"wiiu/tmp/cemu/
-        rsync -avzh "$romsPath"wiiu/tmp/cemu/ "$romsPath"wiiu/
-        rm -rf "$romsPath"wiiu/tmp 
-        rm -f "$romsPath"wiiu/cemu.zip
-	fi
+	setMSG "Installing $Cemu_emuName"		
 
+	curl $Cemu_releaseURL --output "$romsPath"wiiu/cemu.zip 
+	mkdir -p "$romsPath"wiiu/tmp
+	unzip -o "$romsPath"wiiu/cemu.zip -d "$romsPath"wiiu/tmp
+	mv "$romsPath"wiiu/tmp/cemu_*/ "$romsPath"wiiu/tmp/cemu/
+	rsync -avzh "$romsPath"wiiu/tmp/cemu/ "$romsPath"wiiu/
+	rm -rf "$romsPath"wiiu/tmp 
+	rm -f "$romsPath"wiiu/cemu.zip
+	
 	cp "$EMUDECKGIT/tools/launchers/cemu.sh" "${toolsPath}"launchers/cemu.sh
 	sed -i "s|/run/media/mmcblk0p1/Emulation/tools|${toolsPath}|" "${toolsPath}"launchers/cemu.sh
 	sed -i "s|/run/media/mmcblk0p1/Emulation/roms/wiiu|${romsPath}wiiu|" "${toolsPath}"launchers/cemu.sh
@@ -36,7 +32,7 @@ Cemu.install(){
 
 #ApplyInitialSettings
 Cemu.init(){
-	setMSG "Setting up Cemu"	
+	setMSG "Initializing $Cemu_emuName settings."	
 	rsync -avhp $EMUDECKGIT/configs/info.cemu.Cemu/data/cemu/ "${romsPath}wiiu"
     Cemu.setEmulationFolder
 	Cemu.setupSaves
@@ -45,14 +41,16 @@ Cemu.init(){
 
 #update
 Cemu.update(){
-	setMSG "Updating Cemu Config"	
+	setMSG "Updating $Cemu_emuName settings."	
 	Cemu_cemuSettings="${romsPath}wiiu/settings.xml"
     if [ -f $Cemu_cemuSettings ]; then
 	    mv -f $Cemu_cemuSettings $Cemu_cemuSettings.bak #retain cemusettings if it exists to stop wiping peoples mods. Just insert our search path for installed games.
 	fi
     rsync -avhp $EMUDECKGIT/configs/info.cemu.Cemu/data/cemu/ "${romsPath}wiiu"
-	rm $Cemu_cemuSettings
-	mv -f $Cemu_cemuSettings.bak $Cemu_cemuSettings
+	if [ -f $Cemu_cemuSettings.bak ]; then
+	   	rm $Cemu_cemuSettings
+		mv -f $Cemu_cemuSettings.bak $Cemu_cemuSettings
+	fi
     Cemu.setEmulationFolder
 	Cemu.setupSaves
 	Cemu.addSteamInputProfile
@@ -60,7 +58,7 @@ Cemu.update(){
 
 #ConfigurePaths
 Cemu.setEmulationFolder(){
-	setMSG "Setting Cemu Emulation Folder"	
+	setMSG "Setting $Cemu_emuName Emulation Folder"	
     Cemu_cemuSettings="${romsPath}wiiu/settings.xml"
 	if [[ -f "${Cemu_cemuSettings}" ]]; then
 		gamePathEntryFound=$(grep -rnw $Cemu_cemuSettings -e "z:${romsPath}wiiu/roms")
@@ -72,7 +70,6 @@ Cemu.setEmulationFolder(){
 
 #SetupSaves
 Cemu.setupSaves(){
-	setMSG "Linking Cemu Saves to Emulation/saves"
 	unlink "${savesPath}Cemu/saves" # Fix for previous bad symlink
 	linkToSaveFolder Cemu saves "${romsPath}wiiu/mlc01/usr/save"
 }
@@ -94,6 +91,7 @@ Cemu.wipeSettings(){
 
 #Uninstall
 Cemu.uninstall(){
+	setMSG "Uninstalling $Cemu_emuName."
     rm -rf "${Cemu_emuPath}"
 }
 
@@ -133,5 +131,6 @@ Cemu.finalize(){
 }
 
 Cemu.addSteamInputProfile(){
+	setMSG "Adding $Cemu_emuName Steam Input Profile."
 	rsync -r "$EMUDECKGIT/configs/steam-input/cemu_controller_config.vdf" "$HOME/.steam/steam/controller_base/templates/"
 }
